requirements.txt
fastapi>=0.128.8,<0.129.0
uvicorn[standard]>=0.40.0,<0.41.0
pwdlib[argon2]>=0.3.0,<0.4.0
pyjwt>=2.11.0,<3.0.0
sqlalchemy[asyncio]>=2.0.46,<3.0.0
asyncpg>=0.31.0,<0.32.0
alembic>=1.18.4,<2.0.0
pydantic-settings>=2.12.0,<3.0.0
pydantic[email]>=2.12.5,<3.0.0
python-multipart>=0.0.22,<0.0.23
